pyproject.toml
[tool.poetry]
name = "estios"
version = "0.1.0"
description = "Economic Spatial Temporal Input-Output Systems"
authors = ["Dr Griffith Rees <grees@turing.ac.uk>"]
maintainers = ["Dr Griffith Rees <grees@turing.ac.uk>"]
readme = "README.md"
repository = "https://github.com/griff-rees/estios"

[tool.poetry.dependencies]
python = "~3.10"
geopandas = "^0.10.2"
plotly = "^5.3.1"
typer = {version = "^0.4.0", extras = ["all"]}
openpyxl = "^3.0.9"
setuptools = "^62.6.0"
numpy = "^1.23.0"
xlrd = "^2.0.1"
# GDAL = {version = "3.5.3", platform = "darwin"}
# Fiona = "1.8.21"
# pyogrio = "^0.4.2"
# pyogrio = "^0.4.2"
# pyarrow = "^10.0.1"
Fiona = "^1.8.22"
pymrio = "^0.4.8"
country-converter = "^0.8.0"
poetry-dotenv-plugin = "^0.1.0"

[tool.poetry.group.jupyter.dependencies]
jupyterlab = "^3.2.2"

[tool.poetry.group.server]
optional = true

[tool.poetry.group.server.dependencies]
python-dotenv = {version = "^0.19.2", extras = ["cli"]}
jupyter-dash = "^0.4.0"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
dash-auth = "^1.4.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest-remotedata = "^0.3.3"
pytest = "^6.1"
pre-commit = "^2.18.1"
pytest-cov = "^3.0.0"
pytest-sugar = "^0.9.4"
mypy = "^0.982"
pytest-xdist = "^3.1.0"
pymrio = "^0.4.8"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^8.3.9"
mkdocs-typer = "^0.0.2"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}
mkdocs-literate-nav = "^0.5.0"
mkdocs-gen-files = "^0.4.0"
mkdocs-section-index = "^0.3.4"
black = "^22.10.0"
pymdown-extensions = "^9.6"
mkdocs-bibtex = "^2.8.5"
mktestdocs = "^0.2.0"
markdown-exec = {extras = ["ansi"], version = "^1.2.0"}
tabulate = "^0.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = """
--cov=estios
--cov-report=term:skip-covered
--pdbcls=IPython.terminal.debugger:TerminalPdb
--doctest-modules
-m "not slow"
--ignore=notebooks
"""
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
remote_data_strict = true


[tool.poetry.scripts]
estios = "estios.cli:app"

# Pre-commit hooks related config.
[tool.pycln]
all = true
